+++ R/internal.R
@@ -2,6 +2,7 @@
 #' @description Utility functions useful for passing data from functions
 #' inside others.
 #'
+#' @encoding UTF-8
 #' @author Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
 #' @param model_frame a model frame build internally in some functions on
 #'        \strong{EstimationTools} package
@@ -40,8 +41,16 @@ fo_and_data <- function(y, fo, model_frame, data, fo2Surv = TRUE){
       ySurv <- vars[1L]
       yname <- gsub("Surv\\((.*?),.*", "\\1", ySurv)
       statusname <- gsub(paste0("Surv\\(", yname, ",(.*?)\\)"), "\\1", ySurv)
-      factorname <- vars[length(vars)]
-      data <- data.frame(y[,1], y[,2], model_frame[,2])
+      right_hand <- attr(stats::terms(fo), 'term.labels')
+
+      if (length(right_hand) == 0){
+        factorname <- NULL
+        data <- data.frame(y[,1], y[,2])
+      } else {
+        factorname <- as.character(right_hand[1])
+        other_column <- model_frame[,2]
+        data <- data.frame(y[,1], y[,2], other_column)
+      }
       colnames(data) <- c(yname, statusname, factorname)
     }
   }